notes: Fix "out of range" issue on main page
This issue would appear if the user tried to save a note that had any
starting/ending whitespace. The stylecontent prop would have a space in
the array for those too but when saved the content got trimmed and so
those extra whitespaces were removed, causing the toTextSpan method to
throw an error when trying to apply styles. This was resolved by
trimming the stylecontent accordingly.