notes: Rework illustrations
They are not anymore bundled with the app as png and loaded from assets.
Instead we now render those at runtime, but for optimization we cache
the parsed paths so we just get those instead of parsing each time.

Also introduce a generator to convert an svg to a pathinfo object easily